feature(frontend): move send types out of Send file
