+++ frontend/src/components/GddSend/TransactionForm.spec.js
@@ -1,7 +1,7 @@
 import { mount } from '@vue/test-utils'
 import TransactionForm from './TransactionForm'
 import flushPromises from 'flush-promises'
-import { SEND_TYPES } from '@/pages/Send'
+import { SEND_TYPES } from '@/utils/sendTypes'
 import { createMockClient } from 'mock-apollo-client'
 import VueApollo from 'vue-apollo'
 import { user, selectCommunities as selectCommunitiesQuery } from '@/graphql/queries'
