+++ frontend/src/utils/sendTypes.js
@@ -1,0 +1,4 @@
+export const SEND_TYPES = {
+  send: 'send',
+  link: 'link',
+}
